feat: implement getKeypointPatch
